refactor: UserContext 위치를 application 계층으로 리팩토링하여 계층 책임 분리
